i41: Move misc items from /frontend to /src/library/utils
