+++ src/screens/splash/index.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import { Text, TouchableOpacity, Animated, Image, NativeModules, View, StyleSheet} from 'react-native';
-import * as twitterApi from '../../../frontend/@@vendor/twitter';
+import * as twitterApi from '../../library/utils/@@vendor/twitter';
 import config from '../../../config';
 import { colors, shapes } from '../../../frontend/styles'
